requirements.txt
# YOLOv8 依赖
ultralytics==8.0.224
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# 图像处理和数据科学
opencv-python==4.8.0.74
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
tqdm==4.65.0

# MobileNet 依赖
tensorflow==2.13.0
keras==2.13.1

# 数据处理
scikit-learn==1.3.0
pillow==10.0.0

# 其他工具
pyyaml==6.0.1
requests==2.31.0
pycocotools==2.0.7